requirements.txt
numpy
joblib
dataclasses
black
flake8
autopep8
mypy
mlflow
pyyaml
git+https://github.com/rick0000/mjaigym.git@0.2.0
h5py
scikit-learn